Move page components to pages folder
