+++ packages/frontend/src/components/BodyArea/index.tsx
@@ -1,5 +1,5 @@
 import { PropsWithChildren, Suspense } from 'react';
-import LoadingPage from '../LoadingPage';
+import LoadingPage from '../../pages/LoadingPage';
 import './index.css';
 
 export default function BodyArea({ children }: PropsWithChildren<{}>) {
