+++ packages/frontend/src/components/LoggedOutApp/index.tsx
@@ -2,7 +2,7 @@ import { lazy } from 'react';
 import { Route, Switch } from 'react-router-dom';
 import Header from '../Header';
 
-const LoginPage = lazy(() => import('../LoginPage'));
+const LoginPage = lazy(() => import('../../pages/LoginPage'));
 
 export default function LoggedOutApp() {
   return (
