+++ packages/frontend/src/pages/SearchPage/index.css
