+++ packages/frontend/src/pages/SearchPage/index.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useState } from 'react';
 import { LinkDetail, postSearch } from '../../api-types';
 import { getParamAsArray, getParamAsString } from '../../util/getParam';
-import LinkItem from '../LinkItem';
-import SearchForm from '../SearchForm';
+import LinkItem from '../../components/LinkItem';
+import SearchForm from '../../components/SearchForm';
 import './index.css';
 
 export default function SearchPage() {
